Convert `GoalProblem` enum to a 64 bit value (breaks ABI)
We are running out of available slots for problem types in 32 bit
value type.

Since I believe we have already added most of the types we will need 64
types in total should suffice even in the long term future.

+++ include/libdnf5/base/goal_elements.hpp
@@ -92,7 +92,7 @@ enum class ProblemRules {
 };
 
 /// Define a type of information, hint, or problem gathered during libdnf5::Goal::resolve()
-enum class GoalProblem : uint32_t {
+enum class GoalProblem : uint64_t {
     NO_PROBLEM = 0,
     SOLVER_ERROR = (1 << 0),
     NOT_FOUND = (1 << 1),
